requirements.txt
python-telegram-bot==20.6
fastapi==0.110.0
uvicorn==0.29.0
stripe==9.7.0
python-dotenv==1.0.1
requests==2.31.0
